fixed errors

+++ main.cpp
@@ -90,7 +90,7 @@ public:
     }
 
     Entity(const std::string& texture_path, const float& scaleX,
-           const float& scaleY, const float& posX, const float& posY, const float& speed_) :
+           const float& scaleY, const double& posX, const double& posY, const float& speed_) :
             textureRect(0,0,0,0), frameCount(0), facing(RIGHT), speed(speed_) {
 
         if(!image.loadFromFile(texture_path)){
@@ -107,7 +107,7 @@ public:
     }
 
     Entity(const std::string& texture_path, const int frameCount_ , const direction& facing_, const float& scaleX,
-           const float& scaleY, const float& posX, const float& posY, const float& speed_) :
+           const float& scaleY, const double& posX, const double& posY, const float& speed_) :
             frameCount(frameCount_), facing(facing_), speed(speed_) {
         if(!image.loadFromFile(texture_path)){
             std::cout << "Trouble loading image!\n";
@@ -213,7 +213,7 @@ private:
 public:
 
     Player(const int& health_, const std::string& texture_path, const int frameCount, const Entity::direction& facing, const float& scaleX, const float& scaleY,
-           const float& posX, const float& posY, const Weapon::weapon_types weapon_type_,
+           const double& posX, const double& posY, const Weapon::weapon_types weapon_type_,
            const std::string& weapon_texture_path, const float& speed_) :
             health(health_), player(texture_path,frameCount, facing,
                                     3.f, 3.f, posX, posY, speed_),
